+++ src/core/models/interfaces.ts
@@ -33,13 +33,17 @@ export interface ICuiDictionaryItem<T> {
 }
 
 export interface ICuiMutationHandler {
-    handle(): void;
-    refresh(): void;
+    handle(args: any): void;
+    refresh(args: any): void;
     destroy(): void;
 }
 
 export interface CuiCachable {
     refresh(): boolean;
+}
+
+export interface ICuiParsable {
+    parse(val: any): void;
 }
 
 export interface ICui {